Summit Task
1. A brief explanation of what you chose as a unique identifier and why.
I decided to use a concatenation of student id and student number as a unique identifier since there were repeating numbers in each column but no repetition in the concatenated column.

2. A brief explanation of how you QA’d your output
To test my output, first I counted the number of rows to see if it was the sum of the tables 'washington' and 'california'. Then I counted the number of unique ids, using COUNT, to see if there was any repetition. I did the same to RACE and IEP to see if my updates organized it.

3. A brief explanation of any problems you noticed in the input data.
I did not see anything that I would consider a problem in the input data itself. Instead, there were some challenges in standardizing the column's names and some data like TRUE/FALSE as 0/1 as well as standardizing ethnicity.